requirements.txt
cachetools
pytz
requests
nostr
python-dotenv
schedule